Add TS-590SG CAT protocol specification
This was converted to json using a pdf to json tool initially based off
the TS590SG CAT command reference PDF. It didn't work very well and most
of this was gone over by hand. In hindsight json is not the best for
this. I had an idea of generating the command dispatchers etc at compile
time by using this, but it turned out to be a bit more work and complex
thna I wanted to deal with